Remove ember-cli-deploy
No longer using ember-cli-deploy as we're deploying on vercel.
Additionally keeps things simpler if package.json and the build there is what is used.

Also introduced a new/related env to replace the idea with FABRICA_DEPLOY_TARGET
This env variable can be used if we want to switch/determine
the fabrica env.